requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.44
pymysql==1.1.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
stripe==7.13.0
alembic==1.12.0
pydantic>=1.10.20
email-validator==1.3.1
fastapi-mail==1.2.8
python-multipart==0.0.6